Updated to work on 0.9.1
Made sure every script loads on the desktop client. Fixed some elements of additionalvisualsettings.js not loading properly and removed elements not loading from a restart. It should be noted that the scripts still fail to load if opening the save file too fast, best to wait a second or two. Also added new event to infiniteseasonalevents.js and fixed notifications not being cleared on startup.