Auto refresh and quest filtering support
You are now able to filter out quests by clicking on them inside the Quest List (window/modal) around their Description area. Quests are basically filtered by their general type. You will know a quest has been filtered when the quest is cancelled automatically after doing so. Quests that are not filtered will automatically be started and auto completed. When there are no current quests left (that aren't filtered) left, then a quest refresh will be attempted, and it will succeed and or you have enough money for it. From my short testing with this while doing the quick write-up for these changes, it may be a bit buggy, and it's best not to try to filter every quest while you have Auto Quest turned on, as it will rip through your money. Just a fair warning.