Implement Click Attack speed adjusting
This adds in the ability to change and adjust the Click Attack speed from 20 clicks/s up to 1000 clicks/s. Currently adjusting the Click Attack speed will not reflect off of Auto Click DPS, but this will be attempted to be done in another update. Changing the Click Attack speed will work automatically and is easily compatible with things like Auto Dungeon, Auto Gym, etc.