Enhanced Auto Click DPS precision
This fixes the last update up and is more of a hotfix, but this will still detail and explain some things. Up until now, it was believed that the Auto Clicker could click 20 times a second (50ms) delay, which means that it was possible to kill 20 Pokémon per second. This was wrong, and what you will see with this update is that if you are 1 shotting Pokémon and have the Auto Clicker set to a delay of 20 clicks per second (50ms), then you would actually be defeating less Pokémon than you would expect. This would be correct, and this update makes seeing the true values more accurately as possible even with the Click DPS you are pushing out every second. Another thing to note is that setting the Auto Clicker to a much higher delay or making it click more (lowering the bar), then you will notice that you wouldn't be defeating as many Pokémon as you would expect. This is also correct, and it's all made accurate. The most Pokémon you will be able to defeat if you are 1 shotting Pokémon is roughly around 100 Pokémon per second, and if you set the clicks very high, then you will see this fluctuate to see what I mean. The main reason why it's possible to only be defeating around 100 Pokémon roughly is due to technical limitations, and how setInterval's minimum delay would be around 10ms, and setting it any lower wouldn't really make any difference. Theoretically, however, it could be possible to be defeating much more Pokémon, if you can break past JavaScript's limitations and its single threaded nature.