Fixes Disable All Notifications
This fixes up, disabling all notifications. This should disable most notifications now, even the few notifications that weren't getting disabled before (like some mining ones off the top of my head iirc). The new method I'm using for this should still keep the more important notifications intact, such as changelogs on an update and downloading saves. If there are any regular notifications that may be important that this disables, it should be easily modifiable to filter specific notifications with this new fix. Also, to stay true to keeping this setting OG, notifications that cause sound will still happen despite the notifications not being shown. The biggest thing that comes with this update here is that no longer do you have to suffer not being able to do specific things like skipping mining layers, as well as other things that show a certain type of popup with this setting being enabled. I believe that's the longest standing bug on this project to date that is now finally fixed. Keep in mind that disabling all notifications also disables Desktop notifications now as well, but this can be changed later on. I mostly just wanted to get this out for the time being so that it can be enjoyed. Thanks to those who may have been waiting for this fix specifically and your continued patience 🙏