Weather cancelling + Complete Weather override
This implements a "Default" option or selection for weather, which will revert Weather back to functioning normally. Also, this changes it so that the changed weather will affect all regions and not just the singular region that you're in. This is so you don't have to keep setting it on game reload when also moving from region to region.