Implement Gyms Button (Quick Gym Select feature)
This adds in a new button to the right of the Dock button.

Clicking on the Gyms button will pull up a list of all the Gyms in your current region that you're in. This includes everything that is classified as a Gym (including Elite Four).

Gyms won't be accessible through this feature unless you've actually unlocked them and can access them via normal means. At least, that's the entire idea of having this feature, so you don't have to go manually navigating the map and being able to have a swift teleport to Gyms (and having it work smoothly with Auto Gym as well).

Also, I did some brief testing with this for the currently unreleased regions (Alola and Galar), both of these should work and should be supported with this new feature.