Implement Quick Dungeons Select (Dungeons Button)
This adds in a Dungeons button, similar to the Gyms button that was implemented not too long ago.

Works pretty much similarly as it.

It will load all the Dungeons in your current region, indicate whether they're unlocked or not, if you can afford to enter them, and also shows a nice overview of the amount of clears easily. You won't be able to go into a dungeon unless it's unlocked, and you can afford to enter them, basically what you'd expect.

Should be supported up to Alola and should be future-proof as well.