Check for adding new Pokémon & remove window usage
The use of window is what I suspect is what's breaking this on Pokeclicker Desktop, so hopefully this finally fixes this 🤞

Also, this adds a check for checking if new Pokemon are basically added to the game and modifying the stored array for saved filter settings. Prior to this, most of the Catch Filter would have been broken due to the new update if you were already using the script and hadn't cleared localStorage.