Implement Pokérus Mode (legal Pokérus spreading)
Implements this suggestion (Issue https://github.com/Ephenia/Pokeclicker-Scripts/issues/165)

There's a new button called "PKRS Mode" inside the Daycare Modal (where the other previous settings are located).

When active, this will place your Starter inside the Hatchery when it is level 100 and ready to breed as soon as possible. Not only that, but Pokémon that do not have Pokérus yet will be placed inside the Hatchery and be prioritized while your Starter remains inside the Hatchery.

There's also a "PKRS Strict", which, when active, will only focus on putting your Starter in the Hatchery and that only. This means that your Hatchery would remain empty while this is active, and no eggs will be placed inside unless your Starter is inside of it, and you have Pokémon available that do not have Pokérus yet.

This should be slightly more efficient with ensuring all Pokémon get Pokérus sooner than if you don't use it with all factors considered.

"PKRS Strict" also would only work when "PKRS Mode" is active.