Implements Karma Mode
This adds in one of my own ideas that I thought could be cool, and something people may find fun to play around with. This is also to help hopefully make up for an oopsie on my behalf with the 1.0 release of the script.

What Karma Mode essentially does, when enabled, is that each time you encounter a Pokémon in a category (method), the base shiny odds will decrease by 1 for that particular category/method. Karma Mode uses the base odds defined by the game itself for each category/method. When you find a shiny for a specific category/method, then the rates will reset for that particular category/method. Essentially, the point of Karma Mode is to make for a karma system that will help guarantee you shiny Pokémon after so many encounters without one. Shiny Bonus also works in Karma Mode and works normally. You can re-open the Shiny Modifier to update and see the odds of anything in particular when Karma Mode is active. While in Karma Mode, some things in the Shiny Modifier will not be able to be edited/modified.

When toggling Karma Mode, the Shiny Modifier modal will close (this is intentional design for now).