Implements Treasure Hunter
Implements the suggestion here from this Issue (https://github.com/Ephenia/Pokeclicker-Scripts/issues/200)

It's basically implemented with how they visually represented it and tried to explain it. I also really enjoy suggestions that are visual like that, it helps me out a ton. Throwing that out there, just for the record.

Basically, how this works is, you can select the type of items that you're generally looking for from the dropdown. When you have  a type of items you're specifically looking for selected, then on a fresh layer, surveying the layer will be prioritized first, then afterwards attempting to skip the layer will end up being prioritized 2nd (if you're using the feature).

This means if you're looking for general items, the Survey catches them, then regardless of the items found in the current layer will be ignored and mining will still happen. If what you're looking for is failed to be found by the Survey, then the auto skip layer feature will work how it was before.

If you'd like to disable Treasure Hunter, then set the drop-down to the default option (Selected Item).