Create scriptfixerupper.user.js
This should clear out localStorage without touching your saves and anything related to the game's vanilla settings. It should reset and clear out pretty much all the stuff related to my scripts. So, I suppose you could also call this a cleanup script for all my other scripts if you wish to no longer use them as well.

Essentially the idea of this is to enable it if you have issues with scripts not loading/working or whatever else then reloading the game, and then disabling the script afterwards.

This is really the only best and realistic solution that I can think of, as the update checks I put in scripts recently (due to me changing values of variables that get stored within localStorage by default), people were running into issues with these update checks.