Fix unlimited protein usage
A "recent" update added 2 new vitamin types, Calcium and Carbos.
This updates the listener in the "item bag" part of the "item" menu so each vitamin can be affected by this script.
The pokemon list sorted for protein seems to work in console, but not in here so i used the in game code to get it to work.

I also seperated the bypassProtein function from the "main" script for readability and changed variable name to better reflect the in-game change in vitamin mecanics.

Also, this will most likely not work for the 4rth, 5th [...] vitamin added if more are added. To fix it, just add X new assignment in line 36 (but it would be great to use a better way to get this to work, this selector is shit).

Line 43 and 44 are only used if the user doesn't click on the vitamin img element but on the text or the outer box.